@@ -1,4 +1,4 @@
-class Attendance < ActiveRecord::Base
+class Attendance < ApplicationRecord
   belongs_to :user
 
   serialize :details, SerializeAttendanceDetail.new
