@@ -1,4 +1,3 @@
-function login_page() {
-  location.href = Routes.login_index_path();
-  return false;
-}
+# Place all the behaviors and hooks related to the matching controller here.
+# All this logic will automatically be available in application.js.
+# You can use CoffeeScript in this file: http://coffeescript.org/
