@@ -1,5 +1,5 @@
 FactoryGirl.define do
   factory :attendance do
-
+    
   end
 end
