@@ -1,5 +1,5 @@
 FactoryGirl.define do
   factory :leave_type do
-
+    
   end
 end
