@@ -1,5 +1,5 @@
 FactoryGirl.define do
   factory :reason do
-
+    
   end
 end
