@@ -1,5 +1,5 @@
 FactoryGirl.define do
   factory :system_config do
-
+    
   end
 end
