@@ -1,5 +1,5 @@
 FactoryGirl.define do
   factory :working_report do
-
+    
   end
 end
